New register flow
Now when a user registers it first sings in via github and if that action succeeds then the desktop application sends a post resquest to register the user in the service. There is still a bug identified with this, when the user registers there needs to be a way to update its username on the database or else he wont be able to login.

Only previous authorized emails can now register as a teacher, this is enforced by a file managed by the admin.